+++ ffi/bindings/c/master_example.c
@@ -274,7 +274,7 @@ int main()
         .on_change = &client_state_on_change,
         .ctx = NULL,
     };
-    master_t* master = runtime_add_master_tcp(        
+    master_t* master = master_create_tcp_session(        
         runtime,
         LinkErrorMode_Close,
         master_config,
