+++ ffi/dnp3-schema/src/lib.rs
@@ -216,13 +216,12 @@ pub fn build_lib() -> Result<Library, BindingError> {
     // Master stuff
     let (request, variation_enum) = request::define(&mut builder)?;
     let read_handler = handler::define(&mut builder, variation_enum, &shared_def)?;
-    let (runtime, master_class) = runtime::define(&mut builder, &shared_def)?;
-    let association_class = master::define(&mut builder, &shared_def, master_class, read_handler)?;
+    let association_class = master::define(&mut builder, &shared_def, read_handler)?;
     association::define(&mut builder, association_class, request, &shared_def)?;
 
     // Outstation stuff
     let database = database::define(&mut builder, &shared_def)?;
-    outstation::define(&mut builder, runtime, database, &shared_def)?;
+    outstation::define(&mut builder, database, &shared_def)?;
 
     builder.build()
 }
