+++ ffi/dnp3-schema/src/runtime.rs
@@ -1,7 +1,3 @@
-use std::time::Duration;
-
-use crate::shared::SharedDefinitions;
-use class::ClassHandle;
 use oo_bindgen::class::ClassDeclarationHandle;
 use oo_bindgen::native_function::*;
 use oo_bindgen::native_struct::*;
@@ -9,8 +5,7 @@ use oo_bindgen::*;
 
 pub fn define(
     lib: &mut LibraryBuilder,
-    shared: &SharedDefinitions,
-) -> std::result::Result<(ClassDeclarationHandle, ClassDeclarationHandle), BindingError> {
+) -> std::result::Result<ClassDeclarationHandle, BindingError> {
     // Forward declare the class
     let runtime_class = lib.declare_class("Runtime")?;
 
@@ -37,7 +32,7 @@ pub fn define(
         )?
         .return_type(ReturnType::new(
             Type::ClassRef(runtime_class.clone()),
-            "Handle to the created runtime, {null} if an error occured",
+            "Handle to the created runtime, {null} if an error occurred",
         ))?
         .doc(
             doc("Creates a new runtime for running the protocol stack.")
@@ -55,135 +50,13 @@ pub fn define(
         )?
         .build()?;
 
-    let client_state_enum = lib
-        .define_native_enum("ClientState")?
-        .push(
-            "Connecting",
-            "Client is trying to establish a connection to the remote device",
-        )?
-        .push("Connected", "Client is connected to the remote device")?
-        .push(
-            "WaitAfterFailedConnect",
-            "Failed to establish a connection, waiting before retrying",
-        )?
-        .push(
-            "WaitAfterDisconnect",
-            "Client was disconnected, waiting before retrying",
-        )?
-        .push("Shutdown", "Client is shutting down")?
-        .doc(
-            doc("State of the client connection.")
-                .details("Use by the {interface:ClientStateListener}."),
-        )?
-        .build()?;
-
-    let client_state_listener = lib
-        .define_interface(
-            "ClientStateListener",
-            doc("Callback for monitoring the client connection state")
-                .details("This is registered at creation in {class:Runtime.AddMasterTcp()}."),
-        )?
-        .callback("on_change", "Called when the client state changed")?
-        .param("state", Type::Enum(client_state_enum), "New state")?
-        .return_type(ReturnType::void())?
-        .build()?
-        .destroy_callback("on_destroy")?
-        .build()?;
-
-    let master_config = lib.declare_native_struct("MasterConfiguration")?;
-    let master_config = lib.define_native_struct(&master_config)?
-        .add("address", Type::Uint16, "Local DNP3 data-link address")?
-        .add("level", Type::Enum(shared.decode_log_level.clone()), "Decoding log-level for this master. You can modify this later on with {class:Master.SetDecodeLogLevel()}.")?
-        .add("reconnection_strategy", Type::Struct(shared.retry_strategy.clone()), "Reconnection retry strategy to use")?
-        .add("reconnection_delay", StructElementType::Duration(DurationMapping::Milliseconds, Some(Duration::from_millis(0))), doc("Optional reconnection delay when a connection is lost.").details("A value of 0 means no delay."))?
-        .add(
-            "response_timeout",
-            Type::Duration(DurationMapping::Milliseconds),
-            "Timeout for receiving a response"
-        )?
-        .add("tx_buffer_size", StructElementType::Uint16(Some(2048)), doc("TX buffer size").details("Should be at least 249"))?
-        .add("rx_buffer_size", StructElementType::Uint16(Some(2048)), doc("RX buffer size").details("Should be at least 2048"))?
-        .doc("Master configuration")?
-        .build()?;
-
-    let master_class = lib.declare_class("Master")?;
-
-    let endpoint_list = define_endpoint_list(lib)?;
-
-    let add_master_tcp_fn = lib
-        .declare_native_function("runtime_add_master_tcp")?
-        .param("runtime", Type::ClassRef(runtime_class.clone()), "Runtime to use to drive asynchronous operations of the master")?
-        .param("link_error_mode", Type::Enum(shared.link_error_mode.clone()), "Controls how link errors are handled with respect to the TCP session")?
-        .param("config", Type::Struct(master_config.clone()), "Master configuration")?
-        .param("endpoints", Type::ClassRef(endpoint_list.declaration()), "List of IP endpoints.")?
-        .param("listener", Type::Interface(client_state_listener.clone()), "Client connection listener to receive updates on the status of the connection")?
-        .return_type(ReturnType::new(Type::ClassRef(master_class.clone()), "Handle to the master created, {null} if an error occured"))?
-        .doc(
-            doc("Add a master TCP connection")
-            .details("The returned master must be gracefully shutdown with {class:Master.[destructor]} when done.")
-        )?
-        .build()?;
-
-    let add_master_serial_fn = lib
-        .declare_native_function("runtime_add_master_serial")?
-        .param("runtime", Type::ClassRef(runtime_class.clone()), "Runtime to use to drive asynchronous operations of the master")?
-        .param("config", Type::Struct(master_config), "Master configuration")?
-        .param("path", Type::String, "Path to the serial device. Generally /dev/tty0 on Linux and COM1 on Windows.")?
-        .param("serial_params", Type::Struct(shared.serial_port_settings.clone()), "Serial port settings")?
-        .param("listener", Type::Interface(client_state_listener), "Client connection listener to receive updates on the status of the connection")?
-        .return_type(ReturnType::new(Type::ClassRef(master_class.clone()), "Handle to the master created, {null} if an error occured"))?
-        .doc(
-            doc("Add a master serial connection")
-            .details("The returned master must be gracefully shutdown with {class:Master.[destructor]} when done.")
-        )?
-        .build()?;
-
     // Declare the object-oriented class
-    let _runtime_class = lib
+    let runtime_class = lib
         .define_class(&runtime_class)?
         .constructor(&new_fn)?
         .destructor(&destroy_fn)?
-        .method("AddMasterTcp", &add_master_tcp_fn)?
-        .method("AddMasterSerial", &add_master_serial_fn)?
-        .doc("Event-queue based runtime handle")?
+        .doc("Handle to the underlying runtime")?
         .build()?;
 
-    Ok((runtime_class, master_class))
-}
-
-fn define_endpoint_list(lib: &mut LibraryBuilder) -> Result<ClassHandle, BindingError> {
-    let endpoint_list_class = lib.declare_class("EndpointList")?;
-
-    let endpoint_list_new = lib.declare_native_function("endpoint_list_new")?
-        .param("main_endpoint", Type::String, "Main endpoint")?
-        .return_type(ReturnType::new(Type::ClassRef(endpoint_list_class.clone()), "New endpoint list"))?
-        .doc(doc("Create a new list of IP endpoints.").details("You can write IP addresses or DNS names and the port to connect to. e.g. \"127.0.0.1:20000\" or \"dnp3.myorg.com:20000\"."))?
-        .build()?;
-
-    let endpoint_list_destroy = lib
-        .declare_native_function("endpoint_list_destroy")?
-        .param(
-            "list",
-            Type::ClassRef(endpoint_list_class.clone()),
-            "Endpoint list to destroy",
-        )?
-        .return_type(ReturnType::void())?
-        .doc("Delete a previously allocated endpoint list.")?
-        .build()?;
-
-    let endpoint_list_add = lib.declare_native_function("endpoint_list_add")?
-        .param("list", Type::ClassRef(endpoint_list_class.clone()), "Endpoint list to modify")?
-        .param("endpoint", Type::String, "Endpoint to add to the list")?
-        .return_type(ReturnType::void())?
-        .doc(doc(".").details("You can write IP addresses or DNS names and the port to connect to. e.g. \"127.0.0.1:20000\" or \"dnp3.myorg.com:20000\"."))?
-        .build()?;
-
-    let endpoint_list_class = lib.define_class(&endpoint_list_class)?
-        .constructor(&endpoint_list_new)?
-        .destructor(&endpoint_list_destroy)?
-        .method("Add", &endpoint_list_add)?
-        .doc(doc("List of IP endpoints.").details("You can write IP addresses or DNS names and the port to connect to. e.g. \"127.0.0.1:20000\" or \"dnp3.myorg.com:20000\"."))?
-        .build()?;
-
-    Ok(endpoint_list_class)
+    Ok(runtime_class.declaration.clone())
 }
